+++ src/ffbb_api_client_v2/clients/meilisearch_ffbb_client.py
@@ -1,5 +1,6 @@
 from requests_cache import CachedSession
 
+from ..config import MEILISEARCH_BASE_URL
 from ..helpers.http_requests_helper import default_cached_session
 from ..helpers.meilisearch_client_extension import MeilisearchClientExtension
 from ..models.multi_search_query import (
@@ -24,7 +25,7 @@ class MeilisearchFFBBClient(MeilisearchClientExtension):
     def __init__(
         self,
         bearer_token: str,
-        url: str = "https://meilisearch-prod.ffbb.app/",
+        url: str = MEILISEARCH_BASE_URL,
         debug: bool = False,
         cached_session: CachedSession = default_cached_session,
     ):